Create Creating aggregations
Creating aggregations
This chapter uses data gathered by the National UFO Reporting Center. The data is contained in the Incidents table and in this lesson, you will be aggregating values in the DurationSeconds column.

Instructions

Write a T-SQL query which will return the average, minimum, and maximum values of the DurationSeconds column.