Create Calculating the absolute value
Calculating the absolute value
The Shipments table contains some bad data. There was a problem with the scales, and the weights show up as negative numbers. In this exercise, you will write a query to convert all negative weights to positive weights.

Instructions

Write a query that converts all the negative values in the DeliveryWeight column to positive values.