Create Creating and using variables 1
Creating and using variables
In T-SQL, to create a variable you use the DECLARE statement. The variables must have an at sign (@) as their first character. Like most things in T-SQL, variables are not case sensitive. To assign a value to a variable, you can either use the keyword SET or a SELECT statement, then the variable name followed by an equal sign and a value.

Instructions 1/2

Create an integer variable named counter.
Assign the value 20 to this variables