Create Window functions with aggregations (I)
Window functions with aggregations (I)
To familiarize yourself with the window functions, you will work with the Orders table in this chapter. Recall that using OVER(), you can create a window for the entire table. To create partitions using a specific column, you need to use OVER() along with PARTITION BY.

Instructions

Write a T-SQL query that returns the sum of OrderPrice by creating partitions for each TerritoryName.