Create Calculating mode (II)
Calculating mode (II)
In the last exercise, you created a CTE which assigned row numbers to each unique value in OrderPrice. All you need to do now is to find the OrderPrice with the highest row number.

Instructions

Use the CTE ModePrice to return the value of OrderPrice with the highest row number.